Move Ray logger tests into test_loggers.py
Consolidate tests into the existing test file instead of a separate
test_ray_loggers.py. Use top-level imports instead of local imports.
